feat: add authentication protection to /docs/new page
Implemented server-side authentication check that redirects
unauthenticated users to the sign-in page before accessing
the document creation form.
